app/requirements.txt
Django==2.2.5
djangorestframework==3.8.2
gunicorn==19.9.0
celery==4.2.1
kombu==4.2.0
redis==2.10.6
urllib3